+++ wazo_auth/plugins/http/users/schemas.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from xivo.mallow import fields
@@ -8,7 +8,7 @@ from wazo_auth.schemas import BaseSchema
 
 class _BaseUserSchema(BaseSchema):
 
-    username = fields.String(validate=validate.Length(min=1, max=256), required=True)
+    username = fields.String(validate=validate.Length(min=1, max=256))
     firstname = fields.String(missing=None, allow_none=True)
     lastname = fields.String(missing=None, allow_none=True)
     purpose = fields.String(